% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers in the given list
Returns a threshold value for the lower bound and upper bound of the outliers and Plot a violin plot of the observations}
\usage{
detect_outliers(s)
}
\arguments{
\item{s}{list of double that contains the data}
}
\value{
List of double
A Threshold value for the lower bound and upper bound of the outliers
}
\description{
Detect outliers in the given list
Returns a threshold value for the lower bound and upper bound of the outliers and Plot a violin plot of the observations
}
\examples{
detect_outliers(c(1,1,2,3,4,5,6,9,10,13,40))
}
