% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Plot Correlation Matrix}
\usage{
plot_corr(df, features = NULL)
}
\arguments{
\item{df}{The input data frame for plotting.}

\item{features}{A vector of characters that represents the features names. By default, NULL (returns the correlation plot of all numeric features).}
}
\value{
ggplot object, a correlation plot that illustrates the correlation between numeric features interactively.
}
\description{
A function that generates a pearson correlation plot for a list of numeric features in a given data frame.
}
\examples{

numerical_features <- c('Sepal.Length','Sepal.Width','Petal.Width')

plot_corr(iris, features = numerical_features)

}
