% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\name{plot_histograms}
\alias{plot_histograms}
\title{Detect outliers in the given list
Plots histogram given numeric features of the input dfframe, and
plots bar charts for categorical features of the input dfframe}
\usage{
plot_histograms(df, features, facet_columns = 2)
}
\arguments{
\item{df}{dfFrame that to be used for the plot}

\item{features}{List of string feature names}

\item{facet_columns}{Number of columns in Integer type for facet options}
}
\value{
ggplot object consists of categorical and numeric histograms.
}
\description{
Detect outliers in the given list
Plots histogram given numeric features of the input dfframe, and
plots bar charts for categorical features of the input dfframe
}
\examples{
library(palmerpenguins)
df <- penguins
plot_histograms(df, c("species", "flipper_length_mm", "bill_length_mm", "body_mass_g", "island"), 3)
}
